Added Capturing Forwards
Added function that checks for captures for all pawn and if one is possible, it is mandatory to capture instead of preforming normal move. User can then only use pawns that can caputure and then they can only move them on squares where caputre is preformed. If there is a possiblity of another capture, it can be(or rather is forced to) happen immidietly in the same turn.

Currently there is a bug, that when 2 pawn can caputure diffrent pawns, if one preform caputre, other one can also caputre in the same turn, even tho these captures are preformed by different pawns