Win Condition and Promotion
Added so game doesn't crash when a player or an ai is supposed to win.
Also added so pawns on the last row promote. Currently promoted pawns can't move.
Currently there is a bug whene if a player or an ai cannot move but has still pawns on the board, the game will crash due to not having legal moves and game not ending